+++ src/bool.mli
@@ -23,3 +23,14 @@ include Stringable.S with type t := t
    - [to_int true = 1]
    - [to_int false = 0] *)
 val to_int : t -> int
+
+module Non_short_circuiting : sig
+  (** Non-short circuiting and branch-free boolean operators.
+
+      The default versions of these infix operators are short circuiting, which
+      requires branching instructions to implement. The operators below are
+      instead branch-free, and therefore not short-circuiting. *)
+
+  val (&&) : t -> t -> t
+  val (||) : t -> t -> t
+end