+++ src/sequence.mli
@@ -117,10 +117,10 @@ val filteri : 'a t -> f: (int -> 'a -> bool) -> 'a t
 
 val filter : 'a t -> f: ('a -> bool) -> 'a t
 
-(** [merge t1 t2 ~cmp] merges two sorted sequences [t1] and [t2], returning a sorted
-    sequence, all according to [cmp].  If two elements are equal, the one from [t1] is
+(** [merge t1 t2 ~compare] merges two sorted sequences [t1] and [t2], returning a sorted
+    sequence, all according to [compare].  If two elements are equal, the one from [t1] is
     preferred.  The behavior is undefined if the inputs aren't sorted. *)
-val merge : 'a t -> 'a t -> cmp:('a -> 'a -> int) -> 'a t
+val merge : 'a t -> 'a t -> compare:('a -> 'a -> int) -> 'a t
 
 module Merge_with_duplicates_element : sig
   type ('a, 'b) t =
@@ -150,13 +150,13 @@ module Merge_with_duplicates_element : sig
   [@@@end]
 end
 
-(** [merge_with_duplicates_element t1 t2 ~cmp] is like [merge], except that for each
+(** [merge_with_duplicates_element t1 t2 ~compare] is like [merge], except that for each
     element it indicates which input(s) the element comes from, using
     [Merge_with_duplicates_element]. *)
 val merge_with_duplicates
   :  'a t
   -> 'b t
-  -> cmp:('a -> 'b -> int)
+  -> compare:('a -> 'b -> int)
   -> ('a, 'b) Merge_with_duplicates_element.t t
 
 val hd     : 'a t -> 'a option